+++ example/lib/main.dart
@@ -48,7 +48,7 @@ class MyHomePageState extends State<MyHomePage> {
           child: Column(
             children: <Widget>[
               FormBuilder(
-                context,
+                // context,
                 key: _fbKey,
                 autovalidate: true,
                 // readonly: true,